# person-service/openapi/person-service-api.yaml
openapi: 3.0.0

info:
  title: Person Service
  description: Управление пользователями и их связанными сущностями (адрес, индивидуальные данные)
  version: 1.0.0

servers:
  - url: http://localhost:8082/v1

security:
  - bearerAuth: []

paths:
  /individuals/{userUuid}/archive:
    post:
      operationId: archiveIndividual
      summary: Перевод пользователя в статус "ARCHIVED"
      tags:
        - Individuals
      parameters:
        - $ref: "#/components/parameters/UserUuid"
      responses:
        "204":
          description: Пользователь успешно переведен в статус "ARCHIVED"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Пользователь не найден
        "409":
          description: Пользователь уже находится в статусе "ARCHIVED"

  /individuals/{userUuid}/active:
    post:
      operationId: activateIndividual
      summary: Перевод пользователя в статус "ACTIVE"
      tags:
        - Individuals
      parameters:
        - $ref: "#/components/parameters/UserUuid"
      responses:
        "204":
          description: Пользователь успешно переведен в статус "ACTIVE"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Пользователь не найден
        "409":
          description: Пользователь уже находится в статусе "ACTIVE"

  /individuals/{userUuid}/purge:
    post:
      operationId: purgeIndividual
      summary: Удаление данных пользователя из системы
      tags:
        - Individuals
      parameters:
        - $ref: "#/components/parameters/UserUuid"
      responses:
        "204":
          description: Пользователь успешно удален из системы
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Пользователь не найден

  /individuals:
    post:
      operationId: createIndividual
      summary: Создание профиля пользователя в статусе "PENDING"
      tags:
        - Individuals
      description: Создание профиля пользователя в статусе "PENDING"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndividualCreateProfileRequest"
      responses:
        "201":
          description: Профиль физлица успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndividualInfoResponse'
        "400":
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /individuals/{userUuid}:
    get:
      operationId: getIndividual
      summary: Получение информации по профилю физлица
      tags:
        - Individuals
      parameters:
        - $ref: "#/components/parameters/UserUuid"
      responses:
        "200":
          description: Профиль физлица успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndividualInfoResponse'
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Пользователь с указанным userUuid не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Пользователь не найден
                status: NOT_FOUND

    patch:
      operationId: updateIndividual
      summary: Обновление профиля физлица
      tags:
        - Individuals
      parameters:
        - $ref: "#/components/parameters/UserUuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndividualInfoUpdateRequest'
      responses:
        "200":
          description: Профиль физлица успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndividualInfoResponse'
        "400":
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Пользователь с указанным userUuid не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Пользователь не найден
                status: NOT_FOUND



components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    UserUuid:
      name: userUuid
      in: path
      required: true
      description: Уникальный идентификатор пользователя
      schema:
        type: string
        format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000

  responses:
    Unauthorized:
      description: Недействительный токен
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Недействительный токен
            status: UNAUTHORIZED
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Недостаточно прав
            status: FORBIDDEN

  schemas:
    IndividualCreateProfileRequest:
      type: object
      required:
        - passport_number
        - phone_number
        - email
        - first_name
        - last_name
        - secret_key
        - address
      properties:
        passport_number:
          type: string
          description: Номер паспорта
          example: AB123456
        phone_number:
          type: string
          description: Номер телефона
          example: "+79001234567"
        email:
          type: string
          format: email
          description: Email пользователя, используется как логин в Keycloak
          example: john.doe@example.com
        first_name:
          type: string
          maxLength: 32
          description: Имя
          example: John
        last_name:
          type: string
          maxLength: 32
          description: Фамилия
          example: Doe
        secret_key:
          type: string
          maxLength: 32
          description: Секретный ключ для идентификации клиента
          example: my-secret-key
        address:
          type: object
          required:
            - address
            - zip_code
            - city
            - state
            - country
          properties:
            address:
              type: string
              maxLength: 128
              description: Улица и номер дома
              example: ул. Ленина, 123
            zip_code:
              type: string
              maxLength: 32
              description: Почтовый индекс
              example: "101000"
            city:
              type: string
              maxLength: 32
              description: Город
              example: Москва
            state:
              type: string
              maxLength: 32
              description: Штат / регион
              example: Московская область
            country:
              type: object
              required:
                - name
                - alpha2
                - alpha3
              properties:
                name:
                  type: string
                  maxLength: 32
                  description: Название страны
                  example: Россия
                alpha2:
                  type: string
                  minLength: 2
                  maxLength: 2
                  description: Код страны ISO 3166-1 alpha-2
                  example: RU
                alpha3:
                  type: string
                  minLength: 3
                  maxLength: 3
                  description: Код страны ISO 3166-1 alpha-3
                  example: RUS

    IndividualInfoUpdateRequest:
      type: object
      properties:
        passport_number:
          type: string
          description: Номер паспорта
          example: AB123456
        phone_number:
          type: string
          description: Номер телефона
          example: "+79001234567"
        first_name:
          type: string
          maxLength: 32
          description: Имя
          example: John
        last_name:
          type: string
          maxLength: 32
          description: Фамилия
          example: Doe
        secret_key:
          type: string
          maxLength: 32
          description: Секретный ключ для идентификации клиента
          example: my-secret-key
        address:
          type: object
          properties:
            address:
              type: string
              maxLength: 128
              description: Улица и номер дома
              example: ул. Ленина, 123
            zip_code:
              type: string
              maxLength: 32
              description: Почтовый индекс
              example: "101000"
            city:
              type: string
              maxLength: 32
              description: Город
              example: Москва
            state:
              type: string
              maxLength: 32
              description: Штат / регион
              example: Московская область
            country:
              type: object
              required:
                - name
                - alpha2
                - alpha3
              properties:
                name:
                  type: string
                  maxLength: 32
                  description: Название страны
                  example: Россия
                alpha2:
                  type: string
                  minLength: 2
                  maxLength: 2
                  description: Код страны ISO 3166-1 alpha-2
                  example: RU
                alpha3:
                  type: string
                  minLength: 3
                  maxLength: 3
                  description: Код страны ISO 3166-1 alpha-3
                  example: RUS

    IndividualInfoResponse:
      type: object
      required:
        - id
        - user_id
        - passport_number
        - phone_number
        - status
        - created
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор профиля физлица
          example: 770e8400-e29b-41d4-a716-446655440000
        user_id:
          type: string
          format: uuid
          description: Идентификатор связанного пользователя
          example: 550e8400-e29b-41d4-a716-446655440000
        passport_number:
          type: string
          description: Номер паспорта
          example: AB123456
        phone_number:
          type: string
          description: Номер телефона
          example: "+79001234567"
        status:
          type: string
          description: Статус профиля физлица
          example: ACTIVE
        created:
          type: string
          format: date-time
          description: Дата и время создания профиля
          example: 2024-01-01T00:00:00Z
        address:
          type: object
          properties:
            address:
              type: string
              maxLength: 128
              description: Улица и номер дома
              example: ул. Ленина, 123
            zip_code:
              type: string
              maxLength: 32
              description: Почтовый индекс
              example: "101000"
            city:
              type: string
              maxLength: 32
              description: Город
              example: Москва
            state:
              type: string
              maxLength: 32
              description: Штат / регион
              example: Московская область
            country:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 32
                  description: Название страны
                  example: Россия
                alpha2:
                  type: string
                  minLength: 2
                  maxLength: 2
                  description: Код страны ISO 3166-1 alpha-2
                  example: RU
                alpha3:
                  type: string
                  minLength: 3
                  maxLength: 3
                  description: Код страны ISO 3166-1 alpha-3
                  example: RUS

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Ошибка валидации
        status:
          type: string
          example: BAD_REQUEST